fix(lint): converge just lint to green on clippy 1.90 (batch 2)
Workspace-wide behavior-preserving clippy fixes (all clippy's own suggested
rewrites; build + 2198 tests + diag-host verified unchanged):
manual_is_multiple_of/div_ceil, is_none_or, manual_contains (iter().any→
contains), unnecessary_cast, derivable_impls, vec_init_then_push,
type_complexity (2 type aliases), doc_lazy_continuation, redundant_closure,
needless_range_loop/borrow/lifetimes/question_mark, clone_on_copy, len_zero,
single_char_add_str, field_reassign_with_default, new_without_default,
double_must_use. Scoped reasoned allows: too_many_arguments on 6 geometry/
entry-point fns; approx_constant on a hand-computed 32-gon LUT; nav.rs
expect→let-else+unreachable! documenting the stack invariant; expect/unwrap
allows on 5 dedicated test-harness crates under tests/.

just lint now exits 0 — it was red before the whole track too (floating
@stable clippy 1.90 vs the older baseline the code was written against), only
masked because fmt-check failed first.

+++ userspace/nexus-gfx/src/raster/fill.rs
@@ -185,7 +185,7 @@ pub fn fill_gradient_aa(
     let buf = s.buf_mut();
     for py in y..end_y {
         // Per-row gradient interpolation (fixed point, / (h-1)).
-        let t_num = (py - y) as u32;
+        let t_num = py - y;
         let denom = (h - 1).max(1);
         let mut rgba = [0u8; 4];
         for c in 0..4 {
